05_search_agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
    "tavily-python>=0.7.10",
    "chainlit>=1.0.0",
    "beautifulsoup4>=4.13.4",
    "requests>=2.32.4",
]
